+++ include/libpldm/state_set.h
@@ -155,6 +155,13 @@ enum pldm_state_set_operational_stress_status_values {
 	PLDM_STATE_SET_OPERATIONAL_STRESS_STATUS_NORMAL = 1,
 	PLDM_STATE_SET_OPERATIONAL_STRESS_STATUS_STRESSED = 2,
 };
+/* @brief List of states for Operational Fault status (ID 10).
+ */
+enum pldm_state_set_operational_fault_status_values {
+	PLDM_STATE_SET_OPERATIONAL_FAULT_STATUS_NORMAL = 1,
+	PLDM_STATE_SET_OPERATIONAL_FAULT_STATUS_ERROR = 2,
+	PLDM_STATE_SET_OPERATIONAL_FAULT_STATUS_NON_RECOVERABLE_ERROR = 3,
+};
 
 /* @brief List of states for the Operational Running Status state set (ID 11).
  */